requirements.txt
psycopg2==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
